Ceph minimum client on cinder-plugin-ceph-tempest job enable
Since e222cc976918a331bacff150e84069fda8f4960a, it is possible to set
the minimum client version. The goal of this patch is to enable the
*mimic* client version for the current cinder-plugin-ceph-tempest job.

As a result, we will be able to ensure that snapshots can be deleted
when a volume is cloned from them.

Change-Id: I8ffb0c1130b30a19f336a98a9efef05905786af2
(cherry picked from commit 5e2fc9a58d0c7824f81eb43d20eec86eb5f2d880)