DEMC: Add support for trim/discard
This patch adds support for trim/discard to 3 Dell EMC drivers:

- PowerStore
- PowerMax
- PowerFlex

Support is added to report this capability to both Nova, via the
'discard' value in connection properties, as well as internally to
Cinder, via the 'sparse_copy_volume' driver capability.

PowerStore and PowerMax only support thin volumes, so they will always
report the trimming capability, but the PowerFlex also supports thick
volumes, so it will only be returned for thin volumes, and following
vendor's recommendation [1] it will not be returned for volumes that
have snapshots.

SysAdmins can force all volumes to return discard capabilities for Nova
with the ``report_discard_supported`` configuration option.

[1]: https://docs.delltechnologies.com/bundle/READY_NODE_AMS_DG/page/GUID-8E6676C1-425A-4D51-8F59-1A92DE4F6638.html

Change-Id: Ib119fd689545bdb636a4844c6125aec7c3c24b96