Tests: Fix inspur unit test
Because of how stestr processes and filters tests to run there are tests
that are not being run.

The way stestr executes tests is that a parent process first builds a
list of existing tests, then splits it equally into N lists, which are
passed into as many processes via command line (not fork) and this new
process builds the list of existing tests and then uses the list for
filtering.

The problem is that if a test name changes from when the parent builds
the list to when the child builds the list then the test will not be
run.

That is the case for tests decorated with 'ddt.data' where the data is
not a static value but is dynamically built, for example if they use
'time.time()'.

This patch fixes such unit tests in the Inspur driver, where not only
the test is not run, but if will fail if it is forcefully run with:

  $ tox -epy39 -- -n cinder.tests.unit.volume.drivers.inspur.as13000.test_as13000_driver.AS13000DriverTestCase

The test in question are:

  cinder.tests.unit.volume.drivers.inspur.as13000.test_as13000_driver.AS13000DriverTestCase.test__update_volume_stats_1_#####_####
  cinder.tests.unit.volume.drivers.inspur.as13000.test_as13000_driver.AS13000DriverTestCase.test__update_volume_stats_2_#####_####

Change-Id: I3b434f8f4d515b688d70ef7c8ee3fc9aacdc4846