[SVf]: mkhost failure when volume and node are on different iogrp
Attach operation fails when volume and back-end host are on different
iogroup. The assignment of the node IP addresses is done outside the
try block, and the "for ip_data in lsip_resp" loop. Because of this
the loop was not consistent before, so only latest node was detected
instead of all nodes.

Closes-Bug: #1985065
Change-Id: I72074357140e9817eee704b6e9ebfd7f50343235