[SAP] Add new recount_host_stats API
This patch adds a custom API to force
the volume manager to recount all of it's volumes and snapshots
This enables us to update the allocated capacity for a running
cinder service.

This patch has a shared lock with the _report_driver_status call as well
so if the recount_host_stats is working, then the driver stats won't
get reported until it is complete.

Change-Id: I9a954b0a8d7fd5c359a67a81742add1b69597773