Tests: Monkey patch before logs are loaded
Our tests don't do the eventlet monkey patching in the same way that the
Cinder services -backup, scheduler, volume- do.

This patch changes this by doing the monkey patching as soon as
possible, before cinder OVOs are loaded and bring in a number of oslo
libraries.

When copying the initialization part from the Cinder commands a fix for
'current_thread' is also copied.

Change-Id: I6c44d557466dfcb1f6507917eb88959e2ff388df