Tests: Fix test runner never finishing
The tpool.killall method seems to block in some cases when running unit
tests, as observed with the Guru Meditation Reports. This has been
happening primarily on Python 3.9 and 3.10.

Moving the call from the test setUp to the cleanup after stopping
existing looping calls seems to resolve the issue.  At least we've been
unable to reproduce the issue again even when running tests with
'--random'.

We need a way to identify existing looping calls that must be stopped,
tracking them with a decorator on the LoopingCallBase class __init__
method or looking for its instances in the objects tracked by the
Garbage Collector.

Using the GC mechanisms makes tests take 12 times longer to complete, so
this patch creates a new test utils class called InstanceTracker that
allows tracking any class.

We haven't been able to reproduce the test runner freeze situation with
this patch, even when running tests with '--random', but just in case
this happens again a TPOOL_KILLALL_ISSUE flag has been created to allow
easily disabling things if this were to happen again.

Because the tests use greenthreads, debugging blocked tests with GDB is
a nightmare, so the patch also adds the Guru Meditation Reports feature
to the test runner, that way we can do:

 $ kill -USR2 <child_runner_pid>

And see the greenthreads, where in this case we would see one of them
stuck in the tpool.killall() call.

Change-Id: Idb5c428adcc1112046cc81a441c572a2e6bfa80c
(cherry picked from commit eb8409b3a8af41e2203e3b84bc6f02167ceef6a4)