Tests: Fix PowerMax tests
The PowerMax unit tests are currently a nightmare, and it's a miracle
that they pass at all at the gate, as there are hundreds of execution
ordering variations where they will fail.

There are many tests that can fail, but the mostly do 2 things wrong:

- Modify PowerMax own testing global constants

- Forcefully replacing methods with Mocks and not restoring them
  afterwards.

In some cases they modify non PowerMax libraries, so we may see failures
in completely unrelated tests.

Change-Id: I05d7e857c52c98aad12b4aeb8958d3e65fbce774