Ceph backup: Close source file
When doing a full backup the Ceph backup driver doesn't explicitly close
the source IOWrapper, so it will be closed by the Garbage Collector on a
green thread.

This patch changes this behavior and explicitly calls the close method
on a native thread.

This is the right thing to do, but this patch also helps with a unit
tests situation.

There are cases when running unit tests that the tests will fail to
return to the shell and a single "kill -USER2 <child_pid>" command will
unstack it.  This patch together with os-brick change
Ib3b066a7da071b1c2de78a1a4e569676539bd335 fixes the situation.

The issue is caused by the Garbage Collector calling the IOWrapper close
method and this one calling rbd_image.close().

Change-Id: I59bbad2283a1a2ae0e0bc670722468a2fe407666
(cherry picked from commit fc5307582a967a70998cbca60b85167b66d86cfb)