Tests: Fix NFS tests
The NFS and Remotefs unit tests use a mocking of the cinder
Configuration object as the driver configuration.

This makes the usage of self.override_config not work as expected, and a
couple of tests actually use it, so it will only work if the test run
has run a test that forcefully sets the value.

If we have these tests:

  def test1(self):
      self.configuration.option = True
      ...

  def test2(self):
      self.set_override('option', False)
      ...

It will fail if test2 runs before test1.

There are currently 2 tests that use the override_config in the nfs UT
file and where its success depend on the test runnin order:

- test_create_sparsed_volume
- test_create_nonsparsed_volume

This patch removes the mocking and uses a proper Configuration so that
the whole file can use the standard override_config method.

Change-Id: I9325a677376beb44a8c39d9c07d0b48e02a8844f