Tests: Randomize tests
We've had broken unit tests for a long time, and we don't realize
because they are always being executed in the same order, and when they
start failing it's harder to understand WHY it is happening.

This patch forces all our unit tests runs to execute tests in a
randomized order to ensure we find bad tests faster.

Change-Id: I1d3107652b6bb9c9f83a54e307165aaa41f100fb