Update nova microversion for volume backed instance rebuild
The cinder side change for reimaging a volume merged in yoga[1].
However, it assume the nova side change has microversion 2.91
which was correct during that time but since then many new
features have merged in nova causing the nova side microversion
to bump to 2.93 (as per current upstream change[2]).
This patch updates the nova microversion (which we use) to send
'volume-reimaged' external event to nova notifying about
completion of the operation from cinder side.

[1] https://github.com/openstack/cinder/commit/d69e89ea3ba63d305f2a344f6c43b83a74d731fc
[2] https://review.opendev.org/c/openstack/nova/+/830883

Change-Id: I8fae9e2b138545e97cc60e32bbfb527c2e2072ea