[SAP] Mark a pool as down if overcommited
This patch will check the capacity factors of a pool and
check it's virtual_free_capacity left every time the volume
manager gets the volume stats from the backend.  If the
virtual_free_capacity is <= 0 then the pool is overcommited
and will be marked down.

The new SAPPoolDownFilter will check against this pool_state
and if it's down, then filter the pool out of results for
the scheduler.

[SAP] change overcommit log to info

This patch changes the overcommit error log to info.

[SAP] fix default max_oversubscription_ratio to float

When the pool stats are pulled from the volume driver and the
max oversubscription ratio isn't set yet, the max oversubscription ratio
comes from the default config setting, which is a string.  Have to force
that to a float when using it for calculations.

[SAP] fix default max_oversubscription_ratio to float

When the pool stats are pulled from the volume driver and the
max oversubscription ratio is a string, the volume manager has to
force the value to a float.

Change-Id: I3a8549392842d81e1f5ef7d541c6e741de2464d1