[SAP] Mark a pool as down if overcommited
This patch will check the capacity factors of a pool and
check it's virtual_free_capacity left every time the volume
manager gets the volume stats from the backend.  If the
virtual_free_capacity is <= 0 then the pool is overcommited
and will be marked down.

The new SAPPoolDownFilter will check against this pool_state
and if it's down, then filter the pool out of results for
the scheduler.