NFS update volume attachment format during volume snapshot
During a NFS volume snapshot of an attached volume, a QCOW2 snapshot is
created and is made the active volume for the instance. The associated
volume attachment is however not updated, resulting in an unbootable
instance after a stop/start.

This adds an update to the volume format and attachment connection_info
format during the snapshot.

Closes-Bug: #1989514

Change-Id: I3c638dab2f0b5a2128eaeeea533450d7cdfa1577