Dell PowerFlex: Additionnal params for enabling self signed certificates
Initially before the change
https://review.opendev.org/c/openstack/os-brick/+/810419 was merged
to close the bug https://bugs.launchpad.net/os-brick/+bug/1929223,
verify_cert was always set to False which can lead to security issues.
It has been decided through this change that this option can be set
to True or False based upon security requirements. This change
introduced a regression failure as the value set to the option
is not part of connection_properties.

This patch adds additional params during initialization so that it
can be carried over os-brick and get adequate REST API response.

Closes-Bug: 1990136
Change-Id: I0d266a57f68221a3b1740a7376e152bb64cac729