[SAP] Fix rpc call to remote host to update host capacity
This patch fixes a problem with the volume
rcpapi.update_migrate_volume_capacity()
It was always pulling the remote host to call from the
volume.service_topic_queue, which is always the current host entry, not
the new host that the volume is migrating to.  This is how we were
getting overcommited against remote datastores during volume attachments
inducing a migration to another vcenter.  The remote wasn't actually
getting called to update it's allocated_capacity_gb.