RBD: Default rbd_secret_uuid to the cluster FSID
Currently the `rbd_secret_uuid` configuration option has no default
value, but at the same time the RBD driver doesn't complain if it's
missing, and will only fail on the Nova side if there is no
`rbd_secret_uuid` configured in the `[LIBVIRT]` section of `nova.conf`.

Using Cinder's `rbd_secret_uuid` has been the preferred way since the
Ocata release.

Most deployments set the `rbd_secret_uuid` value in libvirt's secret to
the Ceph cluster FSID value, and that's what this patch does when it is
not defined.

It doesn't change how replication locations are configured, and those
will still require the secret uuid to be defined.

Change-Id: I739ae6ae5b4d9b074d610f6a70371b294a3e70f8