[SAP] Optimize vCenter queries in filter_hosts()
Instead of querying the vCenter with multiple queries for each and every
host and cluster, we try to pre-fetch all the relevant properties and
pass them on to the appropriate functions. Those function take the cache
as optional argument, so they still work as before when called without
it from another source.

Additionally, we only fetch "runtime.inMaintenanceMode" and
"runtime.connectionState" instead of fetching all of "runtime". This
reduces the amount of costly XML-parsing in our cinder-volume pod.

Change-Id: I37c0c0e86439e67aea95cd5efaa20377d3121e64