cinder-backup: use the same backup backend host for incremental backups
Incremental backups only work if there's a previous backup to base
it on. With the posix driver, this means there needs to be a previous
backup on the same host where the incremental backup is created.

This patch ensures that an incremental backup is scheduled on the
host that that contains the base backup being used for the increment.
Previously we were relying on luck for this and much of the time an
incremental backup would fail due for want of a base backup.

Closes-bug: 1952805
Change-Id: Id239b4150b1c8e9f4bf32f2ef867fdffbe84f96d