3PAR: Error out if vol cannot be converted to base
Consider volume and snapshots as below:

v1
|
`-- s1
    |
    `-- v2
        |
        `-- s2

User initiated deletion of snapshot s1.
It failed with some vague message.

Initially, it was suspected that ...
While copying volume v2 (sometimes an intermediate step to break
volume dependency), we send a request to clone the volume v2 to new
base volume; and the exception [1] isn't handled properly.

[1] Conflict (HTTP 409) 32 - volume has a child

However, on further investigation it was found that ...
after a new volume v2 (omv-<id>) is created and
when we try to delete old volume v2 (osv-<id>),
at this point the exception [1] is thrown as error.

This is now handled gracefully. Appropriate error is thrown
if the volume (v2) has snapshot (s2).

Co-Authored-By: raghavendrat <raghavendra-uddhav.tilay@hpe.com>
Closes-Bug: #1994521
Change-Id: I5e7fb425c92cdf8c16d5a86a58ca1a52421543d7