[Pure Storage] Add support for 3-site, trisync, replication.
Add new parameters `pure_trisync_enabled` and
`pure_trisync_pg_name`.

When these parameters are used in conjunction with a
volume type where `type <in> trisync` will create a volume
that is simultaneously replicate to two target arrays, one
synchronously and the other asynchronously.

It is required that two `replication_devices` are provided, one
that is sync and one that is async.

Also adds the ability to retype a volume between `sync` and
`trisync` replication types.

Consistency Groups are also supported for `trisync`` volume
types, as well as cloning `trisync` CGs.

These changes have been tested in-house by Pure and confirmed
to work as expected in the master branch for 2023.1.

Implements: blueprint pure-trisync
Change-Id: Idecb1c0421ece87f59818a65f15fcba1f49d940a