Set backup status to error on VolumeNotFound
When creating a backup and it fails we will try
to update the status on the volume but if it does
not exist we will instead fail on that and the
backup will be stuck in creating status which means
cloud admin need to help resetting the status.

This catches the VolumeNotFound exception and ignores
updating the volume status and thus backup will be
set to error and can be deleted by user instead.

Closes-Bug: #1996049
Change-Id: Iafa92ece7f83323af257c1702df5029469b11739