[stable-only] Use os-brick from source in wallaby
We had an issue in os-brick that caused a failure in cinder gate.
The wallaby patch for the fix is merged[1] but wallaby has
transitioned into EM and we can't release os-brick wallaby anymore.

To fix the cinder gate situation (as seen here[2]), this patch adds
os-brick as a required project for the ceph job (which is failing)
so we will fetch os-brick from source and not the released version
hence fixing the gate situation.

[1] https://review.opendev.org/c/openstack/os-brick/+/864479
[2] https://review.opendev.org/c/openstack/cinder/+/864185

Change-Id: I22a894d51f9018f5643e86e67666b6b41e7d310a