Handle external events in extend volume
The support to extend attached volumes is being added to glance
cinder store with change[1].
Currently, cinder sends external events to nova if the volume is
in ``in-use`` state irrespective of the instance uuid value in the
attachment records.
When using glance cinder store, the instance_uuid field is None in
the attachment record and that signifies that the volume is attached
to glance host and not to a nova instance. We should not send any
external events in this case.

With this change, we check if the instance UUIDs are not None and
send external events to nova. If None, we don't send any external
events.

[1] https://review.opendev.org/c/openstack/glance_store/+/868742

Closes-Bug: 2000724
Change-Id: Ia0c8ff77139524e9934ef354b1455ea01b4c31ef