[Pure Storage] Add new array status for replication capability
This patch adds a new field for the `get pools` command to show
what replication capability the backend array currently has.

This is based on the status of current array connections in the
backend array.

Response will be `async`, `sync`. `sync` or `trisync`.

`trisync` implies support for `sync` and `async`.
`sync` implies support for `async`.

Change-Id: I46cbb986ed73335270d6dd4ad197197648b55506