[stable-em only] move grenade jobs to experimental
Given that "upgrades from Extended Maintenance are not supported
within the Community" [0], move the grenade jobs to the experimental
queue for the stable branches in EM status.  Additionally, remove
the integrated-gate-storage template because it includes grenade.
(The other jobs in that template were already configured in check
and gate in our .zuul.yaml, so they continue to be run.)

[0] https://docs.openstack.org/project-team-guide/stable-branches.html#extended-maintenance

Change-Id: Iee29e0e980faff341418a4f3c21a6ab91fe894dc