[train only] update gate jobs
1. Given that "upgrades from Extended Maintenance are not supported
within the Community" [0], move the grenade jobs to the experimental
queue for the stable branches in EM status.  Additionally, remove
the integrated-gate [1] and integrated-gate-storage [2] templates
that include grenade.  The relevant non-grenade jobs in these
templates are either already configured in our .zuul.yaml or are
added in this patch.

2. The pep8 job has been failing in the gate when it runs
check_exec.py.  Looks like the shebang in that file is causing the
system python to be used instead of the testenv python, so remove
the shebang from the file and change the way the file is invoked
by tox.

[0] https://docs.openstack.org/project-team-guide/stable-branches.html#extended-maintenance
[1] https://opendev.org/openstack/openstack-zuul-jobs/src/commit/cb57f26b0db714fb3c70f34d71a6eadef6a4ec43/zuul.d/zuul-legacy-project-templates.yaml#L10-L36
[2] https://opendev.org/openstack/tempest/src/commit/275fdb147a3281f77916f8550de954cf5f6a02c8/zuul.d/integrated-gate.yaml#L458-L482

Change-Id: I6a162d9011302de0df070cfa44386d545a578012