nvmet: Fix setup methods
On ghange Icae9802713867fa148bc041c86beb010086dacc9 we changed from
using the nvmet CLI interface to using it as a Python library.

In that change we incorrectly wrote the ``setup`` methods signature and
they are all missing the ``err_func`` parameter.  It is not failing
because that's on the non-privileged side of things, and then on the
privileged side it forcefully adds the parameter on the call to the
actual library.

This patch adds the missing parameter and handles it on the
non-privileged side.

Change-Id: I615497616d87dfc1683977feafcfbfb9fab8e248