[SAP] Add action_track to format specific logs for easier debugging
This patch adds the new action_track for cinder.  The purpose of this
module is to add a track() method which can be used to log business
logic actions being taken on resources (volumes, snaps, etc) so we can
then use this action logs to discover our most common failures and help
debug/fix them.

For now the only supported output is sending the actions into the
standard cinder logfile.  But we could add another adapter in the
trace() method for sending the action logs to a rabbitmq queue, and/or
a db table directly.

This patch also adds calls throughout cinder code to track specific
actions against volumes, such as create, delete, attach, migrate.

Updated the test_attachments_manager test to use a magicmock instead of
a sentinel as sentinels don't cover unknown attrubutes like magickmock
does.