Pure: Report SAM-2 addressing mode for LUNs
Pure iSCSI and FC drivers use SCSI SAM-2 addressing mode, which means
that LUNs < 256 use peripheral mode (unmodified values) and LUNs >= 256
use flat mode (2 higher bits of the MSB must be 01b).

This is not the standard behavior of os-brick, which defaults to
SAM/transparent mode.

In this patch the Pure storage driver reports it's addressing mode so
that LUNs with value greater than 255 can be connected by os-brick.

Depends-On: If32d054e8f944f162bdc8700d842134a80049877
Closes-Bug: #2006960
Change-Id: I4abfb6de58340c5ea70e7a796326f5d2089c80eb