Follow up: Hitachi: Change option name `hitachi_pool` to `hitachi_pools`
This patch changes option name ``hitachi_pool`` to ``hitachi_pools`` and
sets ``hitachi_pool`` as deprecated name, along review comment[*1]
On the related patch[*2], we've supported MULTIPLE pools,
but option name is still ``hitachi_pool``.

[1] review comment:
https://review.opendev.org/c/openstack/cinder/+/846977/2/cinder/volume/drivers/hitachi/hbsd_common.py#52
[2] Related-Id: I49ac061011293900b04a7a5b90ff5b840521993d

A parent patch is not the related patch, why diffrent is we would like
to avoid re-uploading this patch by rebase when parent patch is merged.

Change-Id: If31edee3c8d51300b01272cf29e31f6ad3e9dc4f