Make PowerMax tests stable
The problem manifests itself as an unstable result of
test_check_force. It fails at Zuul once in a while. But this
seems impossible: there is no randomness or timing component
in this test. The randomness occurs because of cross-test
interference.

The root cause is that apparently someone forgot a deepcopy().

We also make test_check_force not to depend on the global specs,
seems like an obvious oversight. It is not the fix, only tidies
the code.

Change-Id: If1ef9bceae96ef044209c758eee2fd22b44efa49