Hitachi: Fix to use correct pool on secondary storage
This patch fixes to use correct pool number for a secondary storage on
GAD environment[*1].
Even though the option ``hitachi_mirror_pool`` is to set pool number for
secondary storage, the option does not work and a pool number for
primary storage would be used on secondary storage.
The bug should be fixed and delete the risk to be used unexpected pool
number.

[1] GAD is the storage mirroring product on Hitachi storage. The feature
was implied as the patch
 https://review.opendev.org/c/openstack/cinder/+/796170 and was merged
into Antelope.

Closes-Bug: #2011810
Change-Id: I9c37ada5e6af1f3c28ebd5c3c2a8baf2d88d0a96
(cherry picked from commit 2270611507720ba5c05bb6acaf0a5825711ba992)