Fix: PowerMax test with flipping force flag
This patch fixes the error faced in a recent gate run[1].
This is a random behavior but there are other places where
similar change is required.
Since the whole powermax unit tests need to be refactored to do
this correctly, this patch only addresses the failure faced in
the gate run.

[1] https://storage.gra.cloud.ovh.net/v1/AUTH_dcaab5e32b234d56b626f72581e3644c/zuul_opendev_logs_bd9/852654/16/gate/openstack-tox-py310/bd9f0d9/testr_results.html

Change-Id: I6dacbb2aba37dac27ad741aee01cd4f1355198cb