HPE XP and NEC V: Host group name is not correct
Host group[1] name formats, which is managed by cinder driver,
for each storage models must be:
``HBSD-xxx`` for Hitachi storage(``xxx`` is WWPN of target host on FC,
or IP address of target host)
``HPEXP-xxx`` for HPE-XP storage
``NEC-xxx`` for NEC V storage , but the format `HBSD-xxx` is used for
OEM storage models because a bug in the merged patch
https://review.opendev.org/c/openstack/cinder/+/866526 overwrites the
prefix for Hitachi storage.

It should be fixed to use original prefixes.

[1] ``Host group``(or ``iSCSI target``), which is one of Hitachi and
OEM storage parameter,
is a group of multiple server hosts connecting to same storage port.

Following site would be help to know what host group is:
https://knowledge.hitachivantara.com/Documents/Management_Software/SVOS/9.1.x/Volume_Management_-_VSP_G130%2C_G%2F%2FF350%2C_G%2F%2FF370%2C_G%2F%2FF700%2C_G%2F%2FF900/Provisioning/11_About_LUN_Manager%2C_logical_units_(LUs)%2C_and_host_groups

Closes-Bug: #2012515
Change-Id: I1c9677112caa0808dff17cbde2db6afbe25a2129