Ceph: Fix restoring old backups to a different backend
There might be scenarios where a very old backup, created from a Cinder
snapshot, was not handled correctly by my previous patch.

This new patch addresses this corner case by adding a fallback
mechanism.

In Cinder, the base name of the backup volume is generated based either
on the volume id alone or a combination of the volume id and backup id.
Two different scenarios are considered here, each corresponding to a
different base name format:

(1)
If the backup was made from a volume snapshot, the older base name
format is used: volume-<vol-uuid>.backup.base.

(2)
If the backup wasn't made from a snapshot, a newer base name format is
used: volume-<vol-uuid>.backup-<backup-uuid>.

In both cases, the function attempts to restore the volume from a backup
image by looking for it under the base name. If the image is not found
(i.e., rbd.ImageNotFound), an error message is logged, and then the
function attempts to retrieve the backup volume under the alternative
base name. This is done because it's possible that the backup volume was
created using a different format than the one initially tried.

In other words, the check for another base name serves as a fallback
mechanism, which is used when the backup image is not found under the
expected name. The primary reason behind having two different base name
formats is likely due to a change in the naming convention at some point
in the history of the Cinder project. This approach ensures backward
compatibility and makes it possible to restore older backups that were
created before the change.

Change-Id: I074aeb222ca4566b630b5b76529454da4c8b493b
(cherry picked from commit becf45c6cb5660fae59efe9c92ec7ebff0fcbce2)