[Pure Storage] Fix failure in replication failover
A previous patch incorrectly removed a call using a specific
version of API as part of the replication failover process.

The call being made to the backend fails as one of the
parameters is not supported in later API version.

This patch fixes the call to the backend to use the required
API version.

There are no tempest tests for replication failover, hence
why the original patch passed tests.

Closes-Bug: #2028380
Change-Id: Id4f5f16a04220cc880e17ffd69a8303c93a6e8ba