[SAP] FCD Driver
Patches for making the FCD driver work.

Change FCD driver to use volume.name attribute for directory naming, also use it for temp directory.
This change will help ops to find fcd objects on the datastore, as there is no longer any searchable object in vcenter....

[SAP] Fix vmware fcd driver delete_volume on failure

This patch catches the VIMException when an fcd volume isn't found
during a delete_volume.  This is considered a success.

[SAP] Update FCD driver

This patch updates the FCD driver to work in our deployment.

1. Create volumes on the requested datastore from the scheduler.

2. This patch refactors the driver name used in the location information
for a volume metadata.  This also ensures that the migrate_volume
only works on the appropriate driver for the driver_name.
3. FCD changes that enable cross vcenter migration for FCD based volumes.
4. FCD copy image to volume working
5. Add cross vc unattached migration for pre 8.0u3 vc versions
6. Cleanup after fcd delete, we dispose of the empty folder
7. Fixed retype to fcd.
8. Add new SAPFCD filter

This new filter is to help ensure that a cross vcenter migration
lands on the same pool as the originating pool only for a
volume migration operation for the FCD driver.  This will result
in no data being moved as the datastores are actually the same
netapp backed nfs volume.

9. Fixed a call to _get_temp_image_folder

The fcd driver has to find the location of the folder
based on the volume, which contains the host entry, so
this patch renames _get_temp_image_folder in the fcd driver
to _get_temp_image_folder_from_volume to distinguish itself
from the vmdk driver.

10. store provider_location as ds name

This patch updates the FCD driver to store the
provider_location as the datastore name instead
of the datastore moref.  This is to solve a
potential future issue where an nfs datastore is remounted.
When an nfs datastore is remounted the moref id changes, even
with the same datastore name.

This patch also hard codes the fcd driver to create snapshot
as a clone of the original volume.

11. FCD clone while attached

This patch addresses an issue with vcenter failing to clone
an FCD based volume while that volume is attached to a vm.

12. FCD backup volumes working

This patch refactors the initialize_connection a bit between
the parent class vmdk.py and fcd.py to get volume backup
to work

13. [SAP] Got backup restore working for FCD

This patch changes the terminate_connection call in
the volume manager to fetch the volume object instead
of the volume record from the DB.  The FCD driver needs
the volume object to update the provider location
for the restored volume.

This patch also updates the terminate_connection for the
fcd driver to properly move and rename the restored write handle
vmdk to the correct volume folder and vmdk name.

14. fixed unit tests

This patch fixes the unit tests needed for fcd.

Don't create shadow vm on initialize_connection not from backup pod

Fix unit test

Update fcd.py fix initiate_connection

Shadow vm had a duplicate vmdk file, we can pass a parameter to create a diskless backing to create_backing call

Update fcd.py terminate_connection

Delete the shadow vm on read kind of terminate_connection

[SAP] Fix retype to fcd

This patch fixes a minor issue when retyping an vmdk volume to
an fcd volume.  the provider_location for an FCD volume now needs
to the the datastore name in the path instead of the vmware datastore
id.

Change-Id: I8259ebcb42e3a5a0874e00360413974e393c5779