Coerce booleans to integer values in paginate_query
(Grabbed from oslo.db's I4b90f1873 which fixes the same
issue.)

This fixes a failed request when a volume listing is requested
with sorting by bootable:asc and with a marker id specified.

It looks like we are due to replace our paginate_query() method
with the one from oslo.db, as noted in the comments above
the method.

Closes-Bug: #2027532
Change-Id: Idf75e9cc8be6d9f0be320f8109e45b9b9f93dacf