[NetApp] LUN space allocation support
Space allocation is an important NetApp driver specific feature.
This needs to be set when the cinder volume is created.
This is not related to thin/thick provisioning feature of cinder
volumes.It is independent of that. It enables ONTAP to reclaim
space automatically when host deletes data.This helps ONTAP
and host to see the actual space correctly when the host
deletes data.
It also helps to keep a LUN (cinder volume) online when the
LUN (cinder volume) in ontap runs out of space and containing
volume (in ONTAP) cannot automatically grow more space.

User can configure it by using volume type extra spec.
By default Space allocation value is disabled for ONTAP LUN

netapp:space_allocation: "<is> True"  # to enable space allocation
netapp:space_allocation: "<is> False" # to disable space allocation

Blueprint: netapp-space-allocation-support
Change-Id: Ib7072f3093067ecd8ad84e396aaecec8f15c49ba