Dell PowerMax : Disabling the renaming of a volume before deletion
This change is to speed up the deletion of a volume which doesn't
require to rename the volume prior to delete it.
This removal will allow to save 15s in terms of operation time
and avoid unnecessary pausing time during migration.
This affects only version up to Yoga as this was included into a
bigger patch merged in Zed.
This is just a one line removal and doesn't affect other stuff.

Closes-Bug: #2033398
Change-Id: I80088ba4171516d5848d510bbe986fb92a734e38