Fix: Roll back volume status during reimage failure
When we try to reimage a volume, we update the status of
volume to 'downloading'.
We later validate the image metadata (like image is 'active',
image size is less than volume size, etc), and in case the
validation fails, we currently don't revert the volume status
back to original ('available', 'in-use' etc) and volume stays
in 'downloading' state.

This patch fixes this by catching the failure exception and
doing a DB update to restore the volume status back to it's
previous state.

Closes-Bug: #2036994
Change-Id: I05bf29e2a089b06398414b542b655a8083c9a21f